DOP-5071 substitute repos branches collection env var for constant
